move TracedThread and related classes to spi trace pkg
