+++ pinot-core/src/main/java/org/apache/pinot/core/util/trace/ContextBasedTraceCallable.java
@@ -1,0 +1,50 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.pinot.core.util.trace;
+
+import org.apache.pinot.spi.trace.TraceCallable;
+
+
+/**
+ * Extend {@link TraceCallable} to automatically register/un-register itself to/from a request.
+ */
+public abstract class ContextBasedTraceCallable<V> extends TraceCallable<V> {
+  private final TraceContext.TraceEntry _parentTraceEntry;
+
+  /**
+   * If trace is not enabled, parent trace entry will be null.
+   */
+  public ContextBasedTraceCallable() {
+    _parentTraceEntry = TraceContext.getTraceEntry();
+  }
+
+  @Override
+  public void preCall() {
+    if (_parentTraceEntry != null) {
+      TraceContext.registerThreadToRequest(_parentTraceEntry);
+    }
+  }
+
+  @Override
+  public void postCall() {
+    if (_parentTraceEntry != null) {
+      TraceContext.unregisterThreadFromRequest();
+    }
+  }
+}
