+++ pinot-spi/src/main/java/org/apache/pinot/spi/trace/TraceCallable.java
@@ -1,0 +1,48 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.pinot.spi.trace;
+
+import java.util.concurrent.Callable;
+
+
+/**
+ * Wrapper class for {@link Callable} to provide hooks for some preparation and cleanup around calling the job.
+ */
+public abstract class TraceCallable<V> implements Callable<V> {
+
+  @Override
+  public V call()
+      throws Exception {
+    preCall();
+    try {
+      return callJob();
+    } finally {
+      postCall();
+    }
+  }
+
+  protected void preCall() {
+  }
+
+  protected void postCall() {
+  }
+
+  public abstract V callJob()
+      throws Exception;
+}
